+++ src/test/java/oop/exercises/ex26/base/PaymentCalculatorTest.java
@@ -2,7 +2,6 @@ package oop.exercises.ex26.base;
 
 import org.junit.jupiter.api.Test;
 
-import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -12,15 +11,14 @@ class PaymentCalculatorTest {
     void month_calculator_returns_the_correct_value() {
         PaymentCalculator calculator = new PaymentCalculator();
 
-        assertEquals(70, calculator.calculateMonthsUntilPayedOff(
-                BigDecimal.valueOf(5000), BigDecimal.valueOf(12), BigDecimal.valueOf(100)));
+        assertEquals(70, calculator.calculateMonthsUntilPayedOff(5000, 12, 100));
     }
 
     @Test
     void month_calculator_throws_exception_for_infinity() {
         PaymentCalculator calculator = new PaymentCalculator();
 
-        assertThrows(ArithmeticException.class, () -> calculator.calculateMonthsUntilPayedOff(
-                BigDecimal.valueOf(99999), BigDecimal.valueOf(99), BigDecimal.valueOf(1)));
+        assertThrows(ArithmeticException.class, () -> calculator.calculateMonthsUntilPayedOff(99999, 99, 1));
+
     }
 }
